prevent any output if there are no dependencies
This is an unintended consequence of allowing dependencies - if we only need one test case, can we directly encode it into the script? This commit makes an executive decision to say "no": all test cases must come from dependencies; a script by itself will not result in test results.

Or maybe this should become an option for Questions.